imagegen/requirements.txt
torch>=2.2.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pillow>=10.2.0
huggingface_hub>=0.21.0
python-multipart>=0.0.9
